Non-blocking data reading
In this change, the isDataReady method is introduced as a way to poll on
whether data is ready to be ready or not.  On its first invocation, it
requests that temperature be read and records the timestamp.  On future
invocations, it checks whether enough time has passed, and if so, will
read the data.  Upon reading temperature data, it requests pressure data
and vice versa.  When being run in this mode, calling getTemperature or
getPressure will read the previously stored value (and perform any
appropriate calculations on the data).